+++ spine-csharp/src/EventData.cs
@@ -40,7 +40,10 @@ namespace Spine {
 		public int Int { get; set; }
 		public float Float { get; set; }
 		public string @String { get; set; }
+
 		public string AudioPath { get; set; }
+		public float Volume { get; set; }
+		public float Balance { get; set; }
 
 		public EventData (string name) {
 			if (name == null) throw new ArgumentNullException("name", "name cannot be null.");